Fix import/build error by making sure a component is a server component
The <NavigationSidebar> was not marked as a client component but became one implicitly because it was imported in one.
The fetcher did then fail with weird errors that did not lead to this client-server component topic. Passing the server component as children to the client component (now "*Wrapper") preserves his server-components-ness and solves the bugs.